Extract the RemoteRelayModule class into a Python module.
I decided to abandon the project at this state.
I don't need a web interface anymore, so I renamed it and leaved
only a Python class that allows to interact with the board.

Even at this stage it may be useful for someone, so I made it
importable as a Python module and put some basic instructions to
README.md with hope someone will read them.